# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-Dashboard API description.",
        "title": "Swagger E-Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/edashboard",
    "paths": {
...
        "/outpatient/attendances": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/outpatient/clinic": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/outpatient/clinic-popup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sClinicCode",
                        "name": "sClinicCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/outpatient/demographics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outpatient"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/referral/external": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Referral"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/referral/external-popup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Referral"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sDate",
                        "name": "sDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/referral/internal": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Referral"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/referral/internal-popup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Referral"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sDate",
                        "name": "sDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}